[#601] tests 360: merge additional 360 tests into cove_360/tests.py
